Move chartdata interface to chartcontainer
