--- apps/class-solid/src/components/plots/Base.tsx
@@ -1,6 +1,0 @@
-export interface ChartData<T> {
-  label: string;
-  color: string;
-  linestyle: string;
-  data: T[];
-}
