+++ apps/class-solid/src/components/plots/ChartContainer.tsx
@@ -83,3 +83,9 @@ export function useChartContext() {
   }
   return context;
 }
+export interface ChartData<T> {
+  label: string;
+  color: string;
+  linestyle: string;
+  data: T[];
+}
