+++ apps/class-solid/src/components/plots/LinePlot.tsx
@@ -1,7 +1,7 @@
 import * as d3 from "d3";
 import { For, createSignal } from "solid-js";
 import { AxisBottom, AxisLeft, getNiceAxisLimits } from "./Axes";
-import type { ChartData } from "./Base";
+import type { ChartData } from "./ChartContainer";
 import { Chart, ChartContainer, useChartContext } from "./ChartContainer";
 import { Legend } from "./Legend";
